Move node event handlers to separate file to prevent import errors.
